moving isValidIdentifier to common
