+++ packages/acs-ui-common/review/beta/acs-ui-common.api.md
@@ -51,6 +51,9 @@ export interface _IObjectMap<T> {
     [key: string]: T;
 }
 
+// @public
+export const isValidIdentifier: (identifier: CommunicationIdentifier) => boolean;
+
 // @internal
 export const _logEvent: (logger: AzureLogger, event: TelemetryEvent) => void;
